ramparser : Fix unknown symbol issue when extracting modules.
When running print-modulesi and memtracer in ramparser, we are trying to parse unnamed symbols.
so ramparsers couldnt extract those modules.
On the Linux kernel side, we ignore unnamed symbols: they are uninformative
and are inserted at a whim. The same logic also added a Ramparser side.
That helps to fix the issue.

Change-Id: I4a900dcaa1b54f9d0e6a5a669e1a8d1e00d8626c
Signed-off-by: Poovendhan Selvaraj <quic_poovendh@quicinc.com>