minidump: add support to extract minidumps
Incase if Save Minidump to DDR and upload on next mission
mode boot feature is enabled means, ramparser will not be
able to extract the individual minidump info.

This patch add support to extract the individual minidumps
from the single minidump2mem.bin file.

To use this support, script needs to be invoked with the
param "--dump2mem".

Change-Id: I53d59bcf32f0d2fe436e9db0e39fd84e9299b329
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>